# [WIP] Function to convert Poly system into a system of non-redundant irreducible Groebner bases