--- a/sympy/solvers/polysys.py
+++ b/sympy/solvers/polysys.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from collections import defaultdict
 from typing import Any, Optional, Tuple
 from collections.abc import Sequence, Iterable
 
@@ -14,6 +15,7 @@ from sympy.core.exprtools import factor_terms
 from sympy.core.sorting import default_sort_key
 from sympy.logic.boolalg import Boolean
 from sympy.polys import Poly, groebner, roots
+from sympy.polys.densearith import dup_lshift, dup_add
 from sympy.polys.domains import ZZ, QQ, ZZ_I, QQ_I
 from sympy.polys.polytools import parallel_poly_from_expr, sqf_part
 from sympy.polys.polyerrors import (
@@ -845,55 +847,173 @@ def groebner_basis(polys: list[Poly], gens: Optional[Tuple] = None) -> Tuple[lis
     elif domain.is_AlgebraicField:
         basis = groebner_basis_algebraic(polys, gens)
         return basis, [], []
+    elif domain.is_FractionField:
+        basis, eqs, ineqs = groebner_basis_fracfield(polys, gens)
+        return basis, eqs, ineqs
     else:
         raise DomainError(f"Unsupported domain: {domain}")
 
 
-def groebner_basis_zz(polys: list[Poly], gens: Tuple) -> list[Poly]:
+def _dom_lift_gaussian(e, K):
+    return [e.y, e.x]
 
+
+def _dom_lower_gaussian(l, K):
+    if len(l) < 2:
+        l.extend([0] * (2 - len(l)))
+    return K(l[1], l[0])
+
+
+def _dom_lift_algebraic(e, K):
+    return e.to_list()
+
+
+def _dom_lower_algebraic(l, K):
+    return K(l)
+
+
+def _dom_lift_poly(e, K, dom_lift):
+    d = []
+    for monom, coeff in e.terms():
+        elements = dom_lift(coeff)
+        if len(elements) > len(d):
+            d.extend([{} for _ in range(len(elements) - len(d))])
+        for i, c in enumerate(elements[::-1]):
+            d[i][monom] = c
+    return [K.ring.from_dict(d) for d in d[::-1]]
+
+
+def _dom_lower_poly(l, K, dom_lower):
+    d = defaultdict(list)
+    for i, e in enumerate(l[::-1]):
+        for monom, coeff in e.items():
+            c = dup_lshift([coeff], i, K.dom)
+            d[monom] = dup_add(d[monom], c, K.dom)
+    return K.ring.from_dict({monom: dom_lower(c) for monom, c in d.items()})
+
+
+def dom_get_alg_lift(K):
+    if K.is_ZZ_I or K.is_QQ_I:
+        dom = K.dom
+        lift = lambda e: _dom_lift_gaussian(e, K)
+        lower = lambda l: _dom_lower_gaussian(l, K)
+    elif K.is_AlgebraicField:
+        dom = K.dom
+        lift = lambda e: _dom_lift_algebraic(e, K)
+        lower = lambda l: _dom_lower_algebraic(l, K)
+    elif K.is_PolynomialRing:
+        dom_e, lift_e, lower_e = dom_get_alg_lift(K.dom)
+        dom = dom_e[K.symbols]
+        lift = lambda e: _dom_lift_poly(e, dom, lift_e)
+        lower = lambda l: _dom_lower_poly(l, K, lower_e)
+    else:
+        raise NotImplementedError
+    return dom, lift, lower
+
+
+def _lift_poly(poly: Poly, new_gen: Symbol) -> Poly:
+    domain = poly.domain
+    dom, lift, _ = dom_get_alg_lift(domain)
+
+    new_gens = poly.gens + (new_gen,)
+    new_coeffs = {}
+
+    for monom, coeff in poly.rep.to_dict().items():
+        lifted = lift(coeff)
+        for i, c in enumerate(lifted):
+            new_monom = monom + (i,)
+            new_coeffs[new_monom] = c
+
+    return Poly.from_dict(new_coeffs, *new_gens, domain=dom)
+
+
+def _lower_poly(poly: Poly, orig_domain, gen_index: int) -> Poly:
+    _, _, lower = dom_get_alg_lift(orig_domain)
+
+    coeffs = defaultdict(list)
+    for monom, coeff in poly.rep.to_dict().items():
+        base_monom = monom[:gen_index] + monom[gen_index + 1:]
+        deg = monom[gen_index]
+        while len(coeffs[base_monom]) <= deg:
+            coeffs[base_monom].append(0)
+        coeffs[base_monom][deg] = coeff
+
+    new_coeffs = {}
+    for monom, coeff_list in coeffs.items():
+        if not coeff_list:
+            coeff_list = [0, 0]
+        elif len(coeff_list) == 1:
+            coeff_list.append(0)
+        new_coeffs[monom] = lower(coeff_list)
+
+    return Poly.from_dict(new_coeffs, *poly.gens[:gen_index], domain=orig_domain)
+
+
+def groebner_basis_zz(polys: list[Poly], gens: Tuple) -> list[Poly]:
     exprs = [p.as_expr() for p in polys]
     basis = groebner(exprs, gens)
     return [Poly(b, *gens, domain=ZZ) for b in basis]
 
 
 def groebner_basis_qq(polys: list[Poly], gens: Tuple) -> list[Poly]:
-
-    cleared = [p.clear_denoms() for p in polys]
-    polys_zz = [p[1] for p in cleared]
-    basis = groebner_basis_zz(polys_zz, gens)
+    cleared = [p.clear_denoms()[1] for p in polys]
+    basis = groebner_basis_zz(cleared, gens)
     return [p.set_domain(QQ) for p in basis]
 
 
 def groebner_basis_complex(polys: list[Poly], gens: Tuple) -> list[Poly]:
-
     domain = polys[0].domain
     i = Symbol('i')
     new_gens = gens + (i,)
 
-    # Convert to ZZ[i] first
     if domain.is_QQ_I:
-        cleared = [p.clear_denoms() for p in polys]
-        polys = [p[1] for p in cleared]
+        polys = [p.clear_denoms()[1] for p in polys]
 
-    new_polys = [Poly(p.as_expr().subs(S.ImaginaryUnit, i), *new_gens, domain=ZZ) for p in polys]
-    new_polys.append(Poly(i ** 2 + 1, *new_gens, domain=ZZ))
+    lifted_polys = [_lift_poly(p, i) for p in polys]
+    min_poly = Poly(i ** 2 + 1, *new_gens, domain=ZZ)
+    lifted_polys.append(min_poly)
 
-    basis = groebner([p.as_expr() for p in new_polys], new_gens)
+    basis = groebner([p.as_expr() for p in lifted_polys], new_gens)
 
-    # Filter i-free polynomials and convert back
-    return [Poly(b.subs(i, S.ImaginaryUnit), *gens, domain=domain)
-            for b in basis if i not in b.free_symbols]
+    result = []
+    for b in basis:
+        poly = Poly(b, *new_gens, domain=ZZ)
+        if i not in poly.free_symbols:
+            result.append(_lower_poly(poly, domain, len(gens)))
+
+    return result
+
+
+def groebner_basis_algebraic(polys: list[Poly], gens: Tuple) -> list[Poly]:
+    K = polys[0].domain
+    alpha = Dummy('alpha')
+    new_gens = gens + (alpha,)
+
+    cleared_polys = [p.clear_denoms()[1] for p in polys]
+
+    lifted_polys = [_lift_poly(p, alpha) for p in cleared_polys]
+    min_poly = Poly(K.mod.to_list(), alpha).clear_denoms()[1]
+    lifted_polys.append(min_poly)
+
+    basis = groebner([p.as_expr() for p in lifted_polys], new_gens)
+
+    result = []
+    for b in basis:
+        poly = Poly(b, *new_gens, domain=ZZ)
+        if alpha not in poly.free_symbols:
+            result.append(_lower_poly(poly, K, len(gens)))
+
+    return result
+
 
 def groebner_basis_polyring(polys: list[Poly], gens: Tuple) -> Tuple[list[Poly], list[Poly]]:
-
     domain = polys[0].domain
     base_domain = domain.domain
     params = domain.symbols
 
     # Handle QQ base domain
     if base_domain.is_QQ:
-        cleared = [p.clear_denoms() for p in polys]
-        polys = [p[1] for p in cleared]
+        polys = [p.clear_denoms()[1] for p in polys]
         base_domain = ZZ
 
     injected_polys = [p.inject() for p in polys]
@@ -909,29 +1029,53 @@ def groebner_basis_polyring(polys: list[Poly], gens: Tuple) -> Tuple[list[Poly],
         if all(g not in poly.free_symbols for g in gens):
             param_eqs.append(poly)
         else:
-            var_basis.append(poly.eject(*params))
+            # Convert back to original domain
+            var_basis.append(poly.eject(*params).set_domain(domain))
+
+    # Convert param_eqs to original domain too
+    param_eqs = [p.eject(*params).set_domain(domain) for p in param_eqs]
 
     return var_basis, param_eqs
 
 
-def groebner_basis_algebraic(polys: list[Poly], gens: Tuple) -> list[Poly]:
-    K = polys[0].domain
+def groebner_basis_fracfield(polys: list[Poly], gens: Tuple) -> Tuple[list[Poly], list[Poly], list[Poly]]:
+    domain = polys[0].domain
+    d = Symbol('d')
+    new_gens = gens + (d,)
 
-    cleared_polys = []
+    cleared_data = []
+    inequalities = []
     for p in polys:
-        p_cleared, _ = p.clear_denoms()
-        cleared_polys.append(p_cleared)
+        denom, p_cleared = p.clear_denoms(convert=True)
+        cleared_data.append((denom, p_cleared))
+        if denom != 1:
+            inequalities.append(Poly(denom, *new_gens, domain=domain))
 
-    alpha = Dummy('alpha')
-    new_gens = gens + (alpha,)
-
-    min_poly = Poly(K.mod.to_list(), alpha)
-    min_poly_cleared, _ = min_poly.clear_denoms()
-
-    new_polys = [Poly(p.as_expr().subs(K.ext, alpha), *new_gens) for p in cleared_polys]
-    new_polys.append(min_poly_cleared)
+    new_polys = []
+    for denom, p_cleared in cleared_data:
+        new_polys.append(p_cleared)
+        if denom != 1:
+            new_polys.append(Poly(d - denom, *new_gens, domain=domain))
 
     basis = groebner([p.as_expr() for p in new_polys], new_gens)
 
-    return [Poly(b.subs(alpha, K.ext), *gens, domain=K)
-            for b in basis if alpha not in b.free_symbols]
+    var_basis = []
+    equalities = []
+
+    for b in basis:
+        poly = Poly(b, *new_gens)
+        if d not in poly.free_symbols:
+            if all(g not in poly.free_symbols for g in gens):
+                # Convert equations back to original domain
+                equalities.append(Poly(poly.as_expr(), *gens, domain=domain))
+            else:
+                # Convert basis elements back to original domain
+                var_basis.append(Poly(poly.as_expr(), *gens, domain=domain))
+        else:
+            if poly.degree(d) == 1:
+                coeff = -poly.coeff_monomial(1)
+                if coeff != 0:
+                    # Convert inequalities back to original domain
+                    inequalities.append(Poly(coeff, *gens, domain=domain))
+
+    return var_basis, equalities, inequalities