@aasim-maverick 2025-01-08
I'm sorry, I wanted to push a better working version of the helper functions that handle the domains but I'm kind of stuck with a few things so thought it would be better to put out the sketch to be reviewed meanwhile.