@aasim-maverick 2025-01-22
Forgive me because I might sound foolish here asking this, but I have been thinking about the question :

_"Why always reduce the problem to computing a Groebner basis over `ZZ` ?"_

Ideally I should've asked this question way earlier but I rather just blindly jumped into the implementation mindlessly.