@oscarbenjamin 2025-01-22
> "Why always reduce the problem to computing a Groebner basis over `ZZ` ?"

Yes, that's a good question and no it is not foolish at all.

The original algorithms for Groebner bases were designed to work for multivariate polynomials over a field most commonly the rationals (`QQ`). Subsequently more efficient algorithms have been developed that work over rings such as the integers (`ZZ`). Algorithms for Groebner basis calculation over a ring are not implemented in SymPy although it looks as if they are:
```python
In [1]: print(groebner([x, y]))
GroebnerBasis([x, y], x, y, domain='ZZ', order='lex')
```
Internally this calculation actually computes a Groebner basis over `QQ` but then factors out the denominators at the end. You can see the difference with e.g.:
```python
In [2]: print(groebner([2*x-1, y]))
GroebnerBasis([2*x - 1, y], x, y, domain='ZZ', order='lex')

In [3]: print(groebner([2*x-1, y], domain=QQ))
GroebnerBasis([x - 1/2, y], x, y, domain='QQ', order='lex')
```
It might be that it is best to reduce everything to `QQ` then but I would rather target `ZZ` because in principle that should be what `groebner` uses internally.

The domain itself can be a polynomial ring like:
```python
In [5]: print(groebner([a*x-b, b*y - a], [x, y]))
GroebnerBasis([a*x - b, -a + b*y], x, y, domain='ZZ[a,b]', order='lex')
```
This time the calculation is also performed by converting the domain to a field but the field is a field of rational functions:
```python
In [7]: ZZ[a,b].get_field()
Out[7]: ZZ(a,b)
```
Working with these rational functions is slower than working with the polynomial ring because of all the GCD calculations. Those are also what makes `QQ` slower than `ZZ` but with polynomials it is worse because GCD calculations are more expensive. We can get the same basis though by promoting the generators from the domain:
```python
In [9]: print(groebner([a*x-b, b*y - a], [x, y, a, b]))
GroebnerBasis([a*x - b, -a + b*y], x, y, a, b, domain='ZZ', order='lex')
```
This is now computed internally with polynomials that have coefficients in `QQ` rather than `ZZ(a,b)` but ideally in future it would be done with `ZZ`.

Other domains like `QQ_I` are also a lot slower than `QQ` so it makes sense to convert them. Ultimately I want to make it so that SymPy can use python-flint for multivariate polynomials over `QQ` or `ZZ` and then this will absolutely be the fastest way. Since python-flint does not provide polynomials over arbitrary domains I think it will be fastest if we can reduce everything to polynomials over ZZ and QQ.

We should do some measurements to see what is fastest. I am almost certain that it is faster to promote generators when the ground domain is a polynomial ring but I am less sure about some of the other cases.

Also in the case of a polynomial ring there is another important difference:
```python
In [10]: print(groebner([a*x-b, b*y - a, a], [x, y]))
GroebnerBasis([1], x, y, domain='ZZ[a,b]', order='lex')

In [11]: print(groebner([a*x-b, b*y - a, a], [x, y, a, b]))
GroebnerBasis([a, b], x, y, a, b, domain='ZZ', order='lex')
```
The first case here says that there are no solutions which is true generically considering that `a` and `b` are not the zero polynomials in the ring `ZZ[a,b]`. However if we want to be able to handle the different symbolic cases then we might want the second solution which says that there are solutions if `a` and `b` are both zero but in that case `x` and `y` can be anything. This is like the difference between `factor_system` and `factor_system_cond`:
```python
In [15]: factor_system([a*x-b, b*y - a, a], [x, y])
Out[15]: []

In [16]: factor_system_cond([a*x-b, b*y - a, a], [x, y])
Out[16]: [[a⋅x - b, -a + b⋅y, a]]
```
In general then I think that we should promote any symbolic parameters from the ground domain so that we can find the symbolic conditions on the existence of solutions.

Otherwise if we are talking about domains that don't have symbols like `ZZ_I` or `QQ(sqrt(2))` then it is really just a question of speed but I expect that it will always be fastest to convert everything to a Groebner bases over ZZ.