@aasim-maverick 2025-01-25
> We should do some measurements to see what is fastest. I am almost certain that it is faster to promote generators when the ground domain is a polynomial ring but I am less sure about some of the other cases.

I gave a shot at the above and yes as you claim, promoting generators is faster, even though my benchmarking wasn't very comprehensive in my opinion but it was showing enough.

![Figure_1](https://github.com/user-attachments/assets/7d87b004-41fd-452c-ae6f-2837d4061011)

Above matplotlib plot that was the output of my first go at the test, as I said, not comprehensive enough but showing. 


Thank you for explaining this.