@oscarbenjamin 2025-01-25
It is unclear what exactly that plot is comparing. A proper comparson should be something like:
```python
In [24]: p1 = sqrt(2)*x**2 + y**2 + z**2 + 1

In [25]: p2 = 3*(x + y) + z

In [26]: p3 = (x + y)**2 + (y + z) + 2

In [27]: eqs = [p1, p2, p3]

In [28]: %time ok = groebner(eqs, [x, y, z], domain=QQ.algebraic_field(sqrt(2)))
CPU times: user 53.5 ms, sys: 32 μs, total: 53.5 ms
Wall time: 51.8 ms

In [29]: s = symbols('s')

In [30]: p1 = s*x**2 + y**2 + z**2 + 1

In [31]: p2 = 3*(x + y) + z

In [32]: p3 = (x + y)**2 + (y + z) + 2

In [33]: p4 = s**2 - 2

In [34]: eqs = [p1, p2, p3, p4]

In [35]: %time ok = groebner(eqs, [x, y, z, s])
CPU times: user 22.3 ms, sys: 57 μs, total: 22.4 ms
Wall time: 21.3 ms
```
This is computing the same Groebner basis for the same system but in one case using `Q(sqrt(2))` and in the other case using a symbol `s` and the equation `s**2 - 2`.