@oscarbenjamin 2025-01-25
This is a way to randomly generate a system of equations:
```python
In [53]: M = randMatrix(3, 9) + s*randMatrix(3, 9)

In [54]: x, y, z = symbols('x, y, z')

In [55]: monomials = [x, y, z, x*y, x*z, y*z, z**2, y**2, z**2]

In [56]: eqs = list(M*Matrix(monomials))

In [57]: %time ok = groebner(eqs + [s**2 - 2], [x, y, z, s])
CPU times: user 125 ms, sys: 25 μs, total: 125 ms
Wall time: 124 ms

In [58]: eqs2 = [eq.subs(s, sqrt(2)) for eq in eqs]

In [59]: %time ok = groebner(eqs2, [x, y, z], domain=QQ.algebraic_field(sqrt(2)))
CPU times: user 400 ms, sys: 8.02 ms, total: 408 ms
Wall time: 407 ms
```
We should try different degrees of polynomials, different number of variables and so on.