@aasim-maverick 2025-01-25
>  A proper comparson should be something like:

Yes I tried to implement the same thing. What I felt was missing in my first go was making multiple runs since computation time changes in subsequent runs, randomizing the equations in the system and varying degrees as mentioned in your last comment. This is why I felt the test was not comprehensive enough. 

> This is a way to randomly generate a system of equations:

Thanks a lot!!