@aasim-maverick 2025-02-24
Thank you !!
Nice :)

Yes this works reasonably well and as you expected, dominates the original approach in terms of speed in most "realistic" cases from the test suite. Initially I hardcoded some systems(from https://github.com/sympy/sympy/pull/27459#issuecomment-2660839519) in the script to compare but for all of them the alternative approach seemed slow, then I compared using examples from the tests in `solvers/tests/test_polysys.py` for a realistic comparison and gladly the alternative approach was faster in almost every test case.

I used the following script to do so:
```python
import signal
import time
from sympy import groebner, parallel_poly_from_expr, Poly, GroebnerBasis, Expr, symbols
from sympy.solvers.polysys import factor_system_poly, _is_degenerate
from typing import List, Tuple


class TimeoutException(Exception):
    pass


def timeout_handler(signum, frame):
    import traceback
    print("\nTimeout occurred! Here's where it got stuck:\n")
    traceback.print_stack(frame)
    raise KeyboardInterrupt


def get_irreducible_groebner_bases(eqs, gens=None) -> List[GroebnerBasis]:
    polys, opts = parallel_poly_from_expr(eqs, gens)
    domain, orig_gens = opts.domain, opts.gens

    gb = groebner(polys, *orig_gens, domain=domain)

    irreducible_bases = []
    systems_to_process = [list(gb)]
    processed_systems = set()

    while systems_to_process:
        system = systems_to_process.pop()

        system_key = tuple(sorted(str(p) for p in system))
        if system_key in processed_systems:
            continue
        processed_systems.add(system_key)

        factored_output = factor_system_poly(system)
        factored_systems = [sys for sys in factored_output if not _is_degenerate(sys)]

        if not factored_systems:
            irreducible_bases.append(groebner(system, *orig_gens, domain=domain))
            continue

        if len(factored_systems) == 1 and all(p in factored_systems[0] for p in system) and all(
            p in system for p in factored_systems[0]):
            irreducible_bases.append(groebner(system, *orig_gens, domain=domain))
            continue

        for sys in factored_systems:
            if sys:
                new_gb = groebner(sys, *orig_gens, domain=domain)
                systems_to_process.append(list(new_gb))

    return remove_redundant_bases(irreducible_bases, orig_gens, domain)


def remove_redundant_bases(bases: List[GroebnerBasis], gens=None, domain=None) -> List[GroebnerBasis]:
    if not bases:
        return []

    non_redundant = []

    for i, base1 in enumerate(bases):
        is_redundant = False

        for j, base2 in enumerate(bases):
            if i == j:
                continue

            if is_subset(base1, base2):
                is_redundant = True
                break

        if not is_redundant:
            non_redundant.append(base1)

    return non_redundant


def groebner_sqf(polys: list[Poly]) -> GroebnerBasis:
    """Compute a square-free Groebner basis."""
    gb = groebner(polys)
    sqf_parts = [s for p in gb if (s := p.sqf_part()) != p]
    if sqf_parts:
        gb = groebner(list(gb) + sqf_parts)
    return gb


def is_subset(gb1: GroebnerBasis, gb2: GroebnerBasis) -> bool:
    """Check if solutions for gb1 are a subset of solutions for gb2."""
    return all(gb1.reduce(p)[1] == 0 for p in gb2)


def get_irreducible_groebner_bases_alternative(
    eqs: list[Expr], gens: list[Expr] | None = None
) -> list[GroebnerBasis]:
    polys, opts = parallel_poly_from_expr(eqs, gens)

    systems = factor_system_poly(polys)
    systems = [sys for sys in systems if not _is_degenerate(sys)]

    bases: list[GroebnerBasis] = []

    while systems:
        gb = groebner_sqf(systems.pop())

        factored = False

        for i in reversed(range(len(gb))):
            _, factors = gb[i].factor_list()
            if len(factors) > 1 or factors[0][1] > 1:
                for factor, _ in factors:
                    sys = [*gb[:i], factor, *gb[i + 1 :]]
                    systems.append(sys)
                factored = True
                break

        if not factored:
            if not any(is_subset(gb, b) for b in bases):
                bases = [b for b in bases if not is_subset(b, gb)]
                bases.append(gb)

    return bases


def time_function(func, *args) -> Tuple[float, any]:
    """Time a function execution and return the elapsed time and result."""
    start_time = time.time()
    result = func(*args)
    end_time = time.time()
    return end_time - start_time, result


def are_groebner_bases_equivalent(bases1: List[GroebnerBasis], bases2: List[GroebnerBasis]) -> bool:
    """
    Check if two lists of Groebner bases are mathematically equivalent.
    Returns True if they represent the same solution sets.
    """
    if len(bases1) != len(bases2):
        return False

    for gb1 in bases1:
        found_match = False
        for gb2 in bases2:
            if is_subset(gb1, gb2) and is_subset(gb2, gb1):
                found_match = True
                break
        if not found_match:
            return False

    for gb2 in bases2:
        found_match = False
        for gb1 in bases1:
            if is_subset(gb2, gb1) and is_subset(gb1, gb2):
                found_match = True
                break
        if not found_match:
            return False

    return True


def main():
    a, b, c, d, e, f, g = symbols('a b c d e f g')

    gens = [a, b, c, d, e, f, g]
    eqs = [
        1 + a * b + d * e, 1 + a * c + d * f,
        1 + b * c + e * f,
        g - a ** 2 - d ** 2,
        g - b ** 2 - e ** 2,
        g - c ** 2 - f ** 2
    ]

    signal.signal(signal.SIGALRM, timeout_handler)
    signal.alarm(600)  # 10 minute timeout

    try:
        print("Running original implementation...")
        orig_time, orig_result = time_function(get_irreducible_groebner_bases, eqs, gens)
        print(f"Original implementation took {orig_time:.2f} seconds")
        print("Result:", orig_result)
        print("\n" + "="*50 + "\n")

        print("Running alternative implementation...")
        alt_time, alt_result = time_function(get_irreducible_groebner_bases_alternative, eqs, gens)
        print(f"Alternative implementation took {alt_time:.2f} seconds")
        print("Result:", alt_result)

        print("\nSpeed comparison:")
        print(f"The alternative implementation was {orig_time/alt_time:.2f}x {'faster' if alt_time < orig_time else 'slower'} than the original")

        print("\nOutput Equivalency:")
        are_equal = are_groebner_bases_equivalent(orig_result, alt_result)
        print(f"Both implementations give mathematically equivalent results: {are_equal}")

    except KeyboardInterrupt:
        print("Time's up after 10 minutes!")


if __name__ == "__main__":
    main()
```

Although this kind of felt slow and not a good enough way for testing(less statistical, repetitive hardcoding, tests only a single system in a single run) so I thought why not use the examples caught [here](https://github.com/aasim-maverick/Groebner-Calls-in-SymPy-Test-Suite). 
So I have, to begin with, written the following script:
```python
import json
import csv
import time
import signal
import pandas as pd
import matplotlib.pyplot as plt
from pathlib import Path
from sympy import Expr, sympify
from typing import Dict, List, Tuple, Any
from aasim_playground.solvers_performance_tuning.reduce_poly_system_test import (
    get_irreducible_groebner_bases,
    get_irreducible_groebner_bases_alternative,
    are_groebner_bases_equivalent,
    timeout_handler
)


class GroebnerComparison:
    def __init__(self, json_path: str, timeout_seconds: int = 600):
        self.json_path = json_path
        self.timeout_seconds = timeout_seconds
        self.results = []
        signal.signal(signal.SIGALRM, timeout_handler)

    def parse_system(self, sequence: List[str]) -> Tuple[List[Expr], List[Expr]]:
        """Convert string expressions to SymPy expressions and extract symbols."""
        exprs = [sympify(expr) for expr in sequence]

        all_symbols = set()
        for expr in exprs:
            all_symbols.update(expr.free_symbols)

        gens = sorted(list(all_symbols), key=str)

        return exprs, gens

    def process_single_system(self, sequence: List[str], index: int) -> Dict[str, Any]:
        """Process a single polynomial system and return results."""
        result = {
            'index': index,
            'sequence': sequence,
            'original_time': None,
            'alternative_time': None,
            'original_error': None,
            'alternative_error': None,
            'equivalent_results': None,
            'timeout': False
        }

        try:
            eqs, gens = self.parse_system(sequence)

            signal.alarm(self.timeout_seconds)
            try:
                start_time = time.time()
                orig_result = get_irreducible_groebner_bases(eqs, gens)
                result['original_time'] = time.time() - start_time
            except Exception as e:
                result['original_error'] = str(e)

            signal.alarm(self.timeout_seconds)
            try:
                start_time = time.time()
                alt_result = get_irreducible_groebner_bases_alternative(eqs, gens)
                result['alternative_time'] = time.time() - start_time
            except Exception as e:
                result['alternative_error'] = str(e)

            if result['original_time'] is not None and result['alternative_time'] is not None:
                result['equivalent_results'] = are_groebner_bases_equivalent(orig_result, alt_result)

        except KeyboardInterrupt:
            result['timeout'] = True
        except Exception as e:
            result['original_error'] = str(e)
            result['alternative_error'] = str(e)

        return result

    def run_comparison(self):
        """Run comparison on all systems in the JSON file."""
        with open(self.json_path, 'r') as f:
            systems = [json.loads(line) for line in f if line.strip()]

        for idx, system in enumerate(systems):
            result = self.process_single_system(system['sequence'], idx)
            self.results.append(result)

            print(f"Processed system {idx + 1}/{len(systems)}")
            if result['timeout']:
                print(f"  ⚠️ Timeout occurred")
            elif result['original_error'] or result['alternative_error']:
                print(f"  ⚠️ Error occurred in orig/alt approach")
            elif result['equivalent_results'] is False:
                print(f"  ⚠️ Results not equivalent")

    def save_results(self, output_dir: str):
        """Save results to CSV and generate visualizations."""
        output_dir = Path(output_dir)
        output_dir.mkdir(exist_ok=True)

        csv_path = output_dir / 'results.csv'
        with open(csv_path, 'w', newline='') as f:
            writer = csv.DictWriter(f, fieldnames=self.results[0].keys())
            writer.writeheader()
            writer.writerows(self.results)

        df = pd.DataFrame(self.results)

        plt.figure(figsize=(12, 6))
        valid_times = df[
            (df['original_time'].notna()) &
            (df['alternative_time'].notna()) &
            (~df['timeout'])
        ]

        plt.scatter(
            valid_times['original_time'],
            valid_times['alternative_time'],
            alpha=0.5
        )
        max_time = max(
            valid_times['original_time'].max(),
            valid_times['alternative_time'].max()
        )
        plt.plot([0, max_time], [0, max_time], 'r--')
        plt.xlabel('Original Implementation Time (s)')
        plt.ylabel('Alternative Implementation Time (s)')
        plt.title('Timing Comparison')
        plt.savefig(output_dir / 'timing_comparison.png')
        plt.close()

        valid_times['speedup'] = valid_times['original_time'] / valid_times['alternative_time']

        summary = {
            'total_systems': len(self.results),
            'timeouts': sum(r['timeout'] for r in self.results),
            'non_equivalent': sum(
                not r['equivalent_results']
                for r in self.results
                if r['equivalent_results'] is not None
            ),
            'original_failures': sum(
                r['original_error'] is not None
                for r in self.results
            ),
            'alternative_failures': sum(
                r['alternative_error'] is not None
                for r in self.results
            ),
            'mean_speedup': valid_times['speedup'].mean(),
            'median_speedup': valid_times['speedup'].median(),
            'mean_original_time': valid_times['original_time'].mean(),
            'median_original_time': valid_times['original_time'].median(),
            'mean_alternative_time': valid_times['alternative_time'].mean(),
            'median_alternative_time': valid_times['alternative_time'].median()
        }

        with open(output_dir / 'summary.txt', 'w') as f:
            f.write("Summary Statistics\n")
            f.write("=================\n\n")
            f.write("Performance Metrics:\n")
            f.write(f"Mean Original Time: {summary['mean_original_time']:.4f} seconds\n")
            f.write(f"Median Original Time: {summary['median_original_time']:.4f} seconds\n")
            f.write(f"Mean Alternative Time: {summary['mean_alternative_time']:.4f} seconds\n")
            f.write(f"Median Alternative Time: {summary['median_alternative_time']:.4f} seconds\n")
            f.write(f"Mean Speedup: {summary['mean_speedup']:.4f}x\n")
            f.write(f"Median Speedup: {summary['median_speedup']:.4f}x\n\n")
            f.write("Error Statistics:\n")
            f.write(f"Total Systems: {summary['total_systems']}\n")
            f.write(f"Timeouts: {summary['timeouts']}\n")
            f.write(f"Non-equivalent Results: {summary['non_equivalent']}\n")
            f.write(f"Original Implementation Failures: {summary['original_failures']}\n")
            f.write(f"Alternative Implementation Failures: {summary['alternative_failures']}\n")

        non_equiv = df[df['equivalent_results'] == False]
        if len(non_equiv) > 0:
            non_equiv.to_csv(output_dir / 'non_equivalent_cases.csv', index=False)

        orig_failures = df[df['original_error'].notna()]
        if len(orig_failures) > 0:
            orig_failures.to_csv(output_dir / 'original_failures.csv', index=False)

        alt_failures = df[df['alternative_error'].notna()]
        if len(alt_failures) > 0:
            alt_failures.to_csv(output_dir / 'alternative_failures.csv', index=False)


def main():

    json_path = "/home/nasa/Desktop/SymPy/sympy/aasim_playground/catch_groebner_calls.jsonl"
    output_dir = "/home/nasa/Desktop/SymPy/sympy/aasim_playground/"

    comparison = GroebnerComparison(json_path)
    comparison.run_comparison()
    comparison.save_results(output_dir)


if __name__ == "__main__":
    main()
```

It extracts sequences from the caught `groebner` calls, creates systems of polynomial equations using them, and tests both the approaches for that system. 

Now the results of this script provide us with a few takeaways,

First the scatter plot:
![timing_comparison](https://github.com/user-attachments/assets/52a18220-8ef4-41d9-9634-1fb78629ab96)

Most points appear below the trend line(`x = y`) across multiple runs suggesting the alternative implementation is faster for those cases. Outliers exist too, where the alternative approach performs poorly. 

Second is the summary text:
```
Summary Statistics
=================

Performance Metrics:
Mean Original Time: 0.0061 seconds
Median Original Time: 0.0019 seconds
Mean Alternative Time: 0.0084 seconds
Median Alternative Time: 0.0020 seconds
Mean Speedup: 1.0740x
Median Speedup: 0.9626x

Error Statistics:
Total Systems: 597
Timeouts: 0
Non-equivalent Results: 29
Original Implementation Failures: 82
Alternative Implementation Failures: 263
```
Now if you look closely, the textual output suggests the alternative approach is on average slower than the original, which contradicts with what we generally see on the scatter plot. The alternative implementation has a higher mean time, suggesting it is slower on average. The median times are very close, but still, the alternative is slightly slower. Now it's kind of confusing but I think this contradiction is because of way more frequent failures of the alternative approach. Maybe the scatter plot's apparent speedup is overshadowed by extreme outliers and high failure rates in the alternative approach. 

I tracked the failing test cases for both approaches and most of them fail because(I think) of how I wrote `parse_system` since most of those test cases contain `sqrt` and trigonometric functions in them, I'm not sure if the function `parse_system` handles them properly. Also there might be actual issues to address in both approaches too and might not just be a shortcoming of the benchmarking script. 

Another concern is that, these two approaches aren't really "two sides of the same coin", they aren't always equivalent in terms of giving outputs for the same system, when ideally they should since we wanted to compare speeds and not correctness, so still work is needed to determine the mathematical correctness of both of them. I have tracked the systems they gave non-equivalent results for in another `csv` file. I hope the test of equivalency of outputs is robust enough in `are_groebner_bases_equivalent` (let me know if not). 

One more thing that caught my eye, 
```python
...
    while systems:
        gb = groebner_sqf(systems.pop())

        factored = False

        for i in reversed(range(len(gb))):
            _, factors = gb[i].factor_list()
            if len(factors) > 1 or factors[0][1] > 1:
...
```
The line `if len(factors) > 1 or factors[0][1] > 1` is risky isn't it ? I mean `factors` can be empty too, that is a possibility I guess, so the index might go out of bound.

Anyways, in conclusion, 
I think the two approaches need to be checked for mathematical correctness more, I wished they were equivalent but unsurprisingly things aren't so ideal. Maybe I will look into it further why the alternative approach fails for the sequences so frequently, it isn't very clear to me yet (even though I feel the root cause(s) might be staring me right in the face). 
The discrepancy between the textual and graphical outputs is confusing at first,

> then I compared using examples from the tests in `solvers/tests/test_polysys.py` for a realistic comparison and gladly the alternative approach was faster in almost every test case.

but I still feel the alternative approach is faster and hence dominates the original approach. I hope I'm not wrong here. 

Please let me know what you think....